Rename 'inbound' dir to 'handlers'
